FNAF Game Jam - 0.3
Camera update, at first I was able to switch from camera to camera with left click but there's wasn't a way to return to the main camera. but with the updated camera manager, I was able to open a menu for the camera and click around the buttons to switch towards that camera. I also limited my horizontal and vertical by clamping them with specifics parameters.