Out of Service 0.7.2
Such a stupid waste of time, this biggest problem I had was the doors, and button...pressing on the button is suppose close the door, you are suppose to click it once! but nooooooOOOoooo, I thought something went wrong with the code but in reality it was the action map being all weird, I had to remove an interaction from the binding and all is good with the world!! Either way, I'm still testing some scripts for the task giving amelia... I just need some props that I can use before pushing it forward.