Out of Service 0.7.3
My PC was running slow and I fear that I would lose all my progress with the doors, lights and the player. I ran into an issue of the power system, whenever the player reaches 0 in power, they can still "open" cams but wouldn't be able to move so I changed it so that if they press it again, their movement would be restore.