O.O.S V.1.5
Oscar is basically done, the area manager works, he is now facing in the direction you set manually in each area unlike previously were he could look backwards or forward now he can do a 360.  Scourage is done and refactored whilst Oscar is very simple and doesn't attack the player but drains the power and reduce the time needed to be found. Maybe I'll ask if he should attack the player once that time runs out, or keep him the way he is. I do have a concern, should I make it so that he returns back to his original spot on the map once he is found and set a timer that reactivates him... I'll ask.