O.O.S v.1.6
Oscar is now fully finished. He doesn't randomly teleports when playing (the problem was in the start method as he would teleport there), He will give the player chances to find him before heading towards the player's vents, once in the vent he will drain your power until he feels comfortable attacking you. You will never see it coming ( maybe i would make it so that there's a button in the office to shook the vents to prevent Oscar from killing you...but that might be for later.)  So Oscar is now one of the many animatronic that will jumpscare you.