O.O.S v.1.7
Alera will now appear in the office, be shooed away if the player opens their camera. If we so desired can add events that are tied to alera, meaning if we wanted to change posters, flicker the lights or change how the environment looks, we can definitely can and finally after reaching the max number set, she will appear in the office.