More colorful map
Although I got a few issues with the map, it does look a lot better from having no color to lots of color, I wished there was less materials to use and some weren't converted properly and the size, I had to double/triple in size in order to fit the player and the animatronics who might be bigger than the player themselves. I requested that this map should be the free roam and demo night should be in this map but I would like a bigger and scarier map.